+++ docs/releases.rst
@@ -4,6 +4,8 @@ Releases
 Version 1.1, Released ????? 2014
 --------------------------------
 
+* The ``nosetests`` plugin has an optional ``--tap-outdir`` flag to specify
+  where to store ``.tap`` files.
 * **tappy** has backported support for Python 2.6.
 * **tappy** has support for Python 3.2, 3.3, and 3.4.
 * **tappy** has support for PyPy.