+++ docs/releases.rst
@@ -1,9 +1,11 @@
 Releases
 ========
 
-Version 1.1, Released ????? 2014
---------------------------------
+Version 1.1, Released October 23, 2014
+--------------------------------------
 
+* ``TAPTestRunner`` has a ``set_outdir`` method to specify where to store
+  ``.tap`` files.
 * The ``nosetests`` plugin has an optional ``--tap-outdir`` flag to specify
   where to store ``.tap`` files.
 * **tappy** has backported support for Python 2.6.